Choropleth maps refactor
- Remove files and functions copied over from the CC1 dashboard maps code that  are not required for dynamically loading map data
- Genericise getMapData function in chartDefinition.ts to work with map locations specified as geography_uri rather than la_uri. The generic term 'geography_uri' should be aliased in SPARQL queries that fetch the map GeoJson
- Parameterise getGeoJson in geoJsonLoader.ts so it can load GeoJson SPARQL queries for any geographical/output area level
