+++ src/services/map-data/mapDataLoader.ts
@@ -1,0 +1,6 @@
+import { tidyResults } from "./fetchData";
+
+export const getMapData: any = async (query: string) => {
+  const data: any = await tidyResults(query);
+  return data;
+};
